refactor(01-03): rename screnshot.ts to screenshot.ts and fix import
- Renamed packages/webview-ui/src/lib/utils/screnshot.ts to screenshot.ts via git mv (history preserved)
- Updated single consumer import in SchemaVisualizer.tsx line 27
- No content changes; zero behavioral change
- Verified: no stale 'screnshot' references remain; tsc + vite build pass
