+++ packages/webview-ui/src/lib/utils/screenshot.ts
